requirements.txt
langchain==1.1.3
langgraph==1.0.5
langchain-community==0.4.1
langchain-ollama==1.0.1
langchain-chroma

chromadb==1.4.1

fastapi==0.124.4
uvicorn==0.38.0
pydantic==2.12.5
gradio==6.3.0

rank-bm25==0.2.2
sentence-transformers==5.2.0
numpy==2.2.5

pytesseract==0.3.13
pdfplumber==0.11.9
Pillow==12.0.0

python-dotenv==1.2.1
aiofiles==23.2.1
